requirements.txt
django-filter
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-language-field==0.0.3
django-rest-auth==0.9.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.0.2
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
model-mommy==1.6.0
